# services/gateway-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gateway do Microservices Lab em Go (Gin). A UI de teste é esta página (`/docs`).",
        "title": "gateway-go",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/healthz": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ops"
                ],
                "summary": "Liveness e conexão com o Postgres",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StatusBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/jobs": {
            "get": {
                "description": "Atalho para achar `job_id`s e ver o status de cada job sem abrir o banco.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Lista os jobs mais recentes",
                "parameters": [
                    {
                        "maximum": 200,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Máximo de jobs.",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.JobOut"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "description": "Valida o payload contra `contracts/jobs/\u003ctype\u003e.schema.json` e grava o job e a outbox na mesma transação. Quem publica no RabbitMQ é o `outbox-relay`. `target=all` usa fanout (4 workers); os demais, um worker só.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Enfileira um job",
                "parameters": [
                    {
                        "enum": [
                            "io.sleep",
                            "io.fetch_urls",
                            "cpu.pbkdf2",
                            "data.json_transform",
                            "pipeline.fanout"
                        ],
                        "type": "string",
                        "default": "io.sleep",
                        "description": "Tipo do job.",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "all",
                            "celery",
                            "taskiq",
                            "asyncio",
                            "go"
                        ],
                        "type": "string",
                        "default": "all",
                        "description": "Stack que processa o job.",
                        "name": "target",
                        "in": "query"
                    },
                    {
                        "description": "Payload do tipo escolhido (hoje só io.sleep tem schema).",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.JobPayload"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/api.JobAccepted"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/jobs/{job_id}": {
            "get": {
                "description": "Status agregado e resultado de cada worker.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Consulta um job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identificador devolvido pelo POST /jobs.",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.JobOut"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorBody"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/outbox": {
            "get": {
                "description": "Linhas aguardando o relay (`pending`) ou já publicadas (`published`). Útil para conferir se o job saiu do gateway e se o relay está esvaziando a fila.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debug"
                ],
                "summary": "Inspeciona a outbox",
                "parameters": [
                    {
                        "enum": [
                            "pending",
                            "published",
                            "all"
                        ],
                        "type": "string",
                        "default": "pending",
                        "description": "Filtro por estado.",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "maximum": 200,
                        "minimum": 1,
                        "type": "integer",
                        "default": 50,
                        "description": "Máximo de linhas.",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.OutboxEntryOut"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorBody": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "job não encontrado"
                }
            }
        },
        "api.JobAccepted": {
            "type": "object",
            "properties": {
                "job_id": {
                    "type": "string",
                    "example": "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                }
            }
        },
        "api.JobOut": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "job_id": {
                    "type": "string",
                    "example": "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                },
                "origin": {
                    "type": "string",
                    "enum": [
                        "gateway-py",
                        "gateway-go"
                    ]
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.WorkerResultOut"
                    }
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "running",
                        "completed"
                    ]
                },
                "target": {
                    "type": "string",
                    "enum": [
                        "all",
                        "celery",
                        "taskiq",
                        "asyncio",
                        "go"
                    ]
                },
                "type": {
                    "type": "string",
                    "example": "io.sleep"
                }
            }
        },
        "api.JobPayload": {
            "type": "object",
            "properties": {
                "ms": {
                    "description": "Ms é a duração do sleep em milissegundos (0 a 60000) quando type=io.sleep.",
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "api.OutboxEntryOut": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "job_id": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "routing_key": {
                    "type": "string"
                }
            }
        },
        "api.StatusBody": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "api.WorkerResultOut": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "result": {
                    "type": "object"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "succeeded",
                        "failed"
                    ]
                },
                "worker": {
                    "type": "string",
                    "example": "asyncio"
                }
            }
        }
    }
}